+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* `ActivityProcessor` implements IDisposable
+
 ## 0.3.0-beta
 
 Released 2020-07-23
