+++ test/OpenTelemetry.Exporter.Jaeger.Tests/TestActivityProcessor.cs
@@ -73,7 +73,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
 #endif
         }
 
-        public void Dispose()
+        protected override void Dispose(bool isDisposing)
         {
             this.DisposedCalled = true;
         }
