+++ test/OpenTelemetry.Exporter.ZPages.Tests/TestActivityProcessor.cs
@@ -74,7 +74,7 @@ namespace OpenTelemetry.Exporter.ZPages.Tests
 #endif
         }
 
-        public void Dispose()
+        protected override void Dispose(bool isDisposing)
         {
             this.DisposedCalled = true;
         }
